+++ conjurapi/requests_v2.go
@@ -102,7 +102,7 @@ func (c *ClientV2) ListAuthenticatorsRequest() (*http.Request, error) {
 func (c *ClientV2) authenticatorsURL(authenticatorType string, serviceID string) string {
 	// If running against Secrets Manager SaaS, the account is not used in the URL.
 	account := c.config.Account
-	if isConjurCloudURL(c.config.ApplianceURL) {
+	if c.config.IsSaaS() {
 		account = ""
 	}